+++ Rebus.AzureBlobs.Tests/AsyncHelpers.cs
@@ -4,7 +4,7 @@ using System.Runtime.ExceptionServices;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Rebus.AzureBlobs
+namespace Rebus.AzureBlobs.Tests
 {
     static class AsyncHelpers
     {